refactor: deprecate tree component
Tree component we use for pages and instances got overcomplicated
with time and should be reimplemented to move forward with css
variables.
